AP-1806 Refactor remarks generators
There is duplicated code in several remark generator classes.
This commit DRYs up the code by moving some functionality into a
base_checker class which can be inherited from by all other checker
classes, and adds an Exemptable concern which can be shared by
any checkers which can exempt payments from flagging.
