+++ app/services/remark_generators/amount_variation_checker.rb
@@ -1,13 +1,6 @@
 module RemarkGenerators
-  class AmountVariationChecker
-    def self.call(assessment, collection)
-      new(assessment, collection).call
-    end
-
-    def initialize(assessment, collection)
-      @assessment = assessment
-      @collection = collection
-    end
+  class AmountVariationChecker < BaseChecker
+    include Exemptable
 
     def call
       populate_remarks unless unique_amounts || exempt_from_checking
@@ -23,22 +16,6 @@ module RemarkGenerators
       my_remarks = @assessment.remarks
       my_remarks.add(record_type, :amount_variation, @collection.map(&:client_id))
       @assessment.update!(remarks: my_remarks)
-    end
-
-    def record_type
-      @collection.first.class.to_s.underscore.tr('/', '_').to_sym
-    end
-
-    def exempt_from_checking
-      childcare_payment? && childcare_disallowed?
-    end
-
-    def childcare_payment?
-      record_type == :outgoings_childcare
-    end
-
-    def childcare_disallowed?
-      @assessment.disposable_income_summary.childcare.zero?
     end
   end
 end
