+++ app/services/remark_generators/multi_benefit_checker.rb
@@ -1,14 +1,5 @@
 module RemarkGenerators
-  class MultiBenefitChecker
-    def self.call(assessment, collection)
-      new(assessment, collection).call
-    end
-
-    def initialize(assessment, collection)
-      @assessment = assessment
-      @collection = collection
-    end
-
+  class MultiBenefitChecker < BaseChecker
     def call
       populate_remarks if flagged?
     end
@@ -23,10 +14,6 @@ module RemarkGenerators
       my_remarks = @assessment.remarks
       my_remarks.add(record_type, :multi_benefit, @collection.map(&:client_id))
       @assessment.update!(remarks: my_remarks)
-    end
-
-    def record_type
-      @collection.first.class.to_s.underscore.tr('/', '_').to_sym
     end
   end
 end
